react-flask-app/api-backend/requirements.txt
logging
pandas
fastavro
langdetect
flask
flask_cors
kafka-python
requests
bs4
Flask~=2.3.2
pandas
nltk
beautifulsoup4
PyYAML
pymongo
bson